review: перенести логику из сервиса в репозитории, оптимизировать работу с БД

+++ ewm-service/src/main/java/ru/practicum/event/controller/PublicEventController.java
@@ -36,15 +36,15 @@ public class PublicEventController {
 
     @GetMapping
     List<EventShortDto> getEvents(@RequestParam(required = false) String text,
-                                 @RequestParam(required = false) List<Integer> categories,
-                                 @RequestParam(required = false) Boolean paid,
-                                 @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
-                                 @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
-                                 @RequestParam(required = false) Boolean onlyAvailable,
-                                 @RequestParam(required = false) String sort,
-                                 @RequestParam(required = false, defaultValue = "0") Integer from,
-                                 @RequestParam(required = false, defaultValue = "10") Integer size,
-                                 HttpServletRequest request) {
+                                  @RequestParam(required = false) List<Long> categories,
+                                  @RequestParam(required = false) Boolean paid,
+                                  @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
+                                  @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
+                                  @RequestParam(required = false) Boolean onlyAvailable,
+                                  @RequestParam(required = false) String sort,
+                                  @RequestParam(required = false, defaultValue = "0") Integer from,
+                                  @RequestParam(required = false, defaultValue = "10") Integer size,
+                                  HttpServletRequest request) {
         log.info("getEvents");
         List<EventShortDto> result = publicEventService.getEvents(text, categories, paid, rangeStart, rangeEnd,
                 onlyAvailable, sort, from, size);
